Update to 7.2pr2 and bugfixes
