+++ gradle.properties
@@ -1,16 +1,16 @@
 org.gradle.jvmargs=-Xmx2G
 
 # BTA
-bta_version=7.1_01
+bta_version=7.2-pre2
 
 # Loader & Mod Menu
-loader_version=0.15.6-babric.5-bta
+loader_version=0.15.6-babric.6-bta
 mod_menu_version=2.0.5
 
 # HalpLibe
-halplibe_version=3.5.4
+halplibe_version=4.1.3
 
 # Mod
 mod_version=1.0.3
-mod_group=MISSmyIDMods
-mod_name=TrowelMod
+mod_group=missmyidmods
+mod_name=trowelmod
