--- src/main/java/MISSmyIDMods/TrowelMod/Mixin/mixin.java
@@ -1,5 +1,0 @@
-package missmyidmods.trowelmod.Mixin;
-
-public class mixin
-{
-}
