+++ src/main/resources/assets/trowelmod/textures/item/trowel.png
