+++ src/main/resources/fabric.mod.json
@@ -13,7 +13,7 @@
     "sources": ""
   },
 
-	"icon": "icon.png",
+  "icon": "icon.png",
   "license": "CC0-1.0",
 
   "environment": "*",
@@ -21,22 +21,21 @@
     "main": [
       "missmyidmods.trowelmod.TrowelMod"
     ],
-		"beforeGameStart": [
-          "missmyidmods.trowelmod.TrowelMod"
-		],
-		"afterGameStart": [
-          "missmyidmods.trowelmod.TrowelMod"
-		],
-		"recipesReady": [
-          "missmyidmods.trowelmod.TrowelMod"
-		]
-	},
+    "beforeGameStart": [
+      "missmyidmods.trowelmod.TrowelMod"
+    ],
+    "afterGameStart": [
+      "missmyidmods.trowelmod.TrowelMod"
+    ],
+    "recipesReady": [
+      "missmyidmods.trowelmod.TrowelMod"
+    ]
+  },
   "mixins": [
-		"trowelmod.mixins.json"
+    "trowelmod.mixins.json"
   ],
 
   "depends": {
-		"minecraft": "~7.1",
     "fabricloader": ">=0.15.5"
   },
   "suggests": {
