--- src/main/resources/lang/TrowelMod/en_US.lang
@@ -1,2 +1,0 @@
-item.trowelmod.Trowel.name = Trowel
-item.trowelmod.Trowel.desc = Trowel places random block from your hotbar.
