+++ src/main/resources/lang/trowelmod/en_US.lang
@@ -1,2 +1,2 @@
-item.trowelmod.Trowel.name = Trowel
-item.trowelmod.Trowel.desc = Trowel places random block from your hotbar.
+item.trowelmod.trowel.name = Trowel
+item.trowelmod.trowel.desc = Trowel places random block from your hotbar.
