+++ src/main/resources/trowelmod.mixins.json
@@ -1,7 +1,7 @@
 {
     "required": true,
     "minVersion": "0.8",
-    "package": "missmyidmods.trowelmod.Mixin",
+    "package": "MISSmyIDMods.TrowelMod.Mixin",
     "compatibilityLevel": "JAVA_8",
     "mixins": [
     ],
